% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v4.r
\name{cal.FCgene}
\alias{cal.FCgene}
\title{cal.FCgene}
\usage{
cal.FCgene(g, community1, community2)
}
\arguments{
\item{g}{graph}

\item{module1.name}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}

\item{module2.name}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}
}
\value{
data.frame
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
cal.FCgene(graph, 'module_1','module_2')
}
