% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v8.r
\name{cal.FCSgenes}
\alias{cal.FCSgenes}
\title{cal.FCSgenes}
\usage{
cal.FCSgenes(
  g,
  comm.genelist,
  community1n,
  community2n,
  random,
  ratio,
  randomMethod,
  nCore
)
}
\arguments{
\item{g}{graph}

\item{community1n}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}

\item{community2n}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}
}
\value{
data.frame
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
cal.FCSgenes(g, comm.genelist, community1n, community2n,random,ratio,randomMethod, nCore)
}
