% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructNetwork.r
\name{genelist.overlap}
\alias{genelist.overlap}
\title{genelist.overlap}
\usage{
genelist.overlap(genelist, overlap_filtering = TRUE)
}
\arguments{
\item{genelist}{list of community genes}

\item{overlap_filtering}{TRUE or FALSE}
}
\value{
genelist
}
\description{
get rid of overlapped genes from the community gene list
}
