% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistFunction.r
\name{get.closest.dist}
\alias{get.closest.dist}
\title{get.closest.dist
average of sum of min of every community node based shortest path length}
\usage{
get.closest.dist(distm, xx1, yy1, avg.dist = NULL)
}
\description{
get.closest.dist
average of sum of min of every community node based shortest path length
}
