% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistFunction.r
\name{get.kernel.dist}
\alias{get.kernel.dist}
\title{get.kernel.dist
average of kernelized distance between communities}
\usage{
get.kernel.dist(distm, x, y, avg.dist = NULL)
}
\description{
get.kernel.dist
average of kernelized distance between communities
}
