% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistFunction.r
\name{get.s2b.dist}
\alias{get.s2b.dist}
\title{get.s2b.dist
average of sum of min of every community node based shortest path length (< avg)}
\usage{
get.s2b.dist(distm, xx1, yy1, avg.dist)
}
\description{
get.s2b.dist
average of sum of min of every community node based shortest path length (< avg)
}
