% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistFunction.r
\name{get.separation.dist}
\alias{get.separation.dist}
\title{get.separation.dist
difference between shortest distance of communities and the average shortest distance of each community}
\usage{
get.separation.dist(distm, x, y, avg.dist = NULL)
}
\description{
get.separation.dist
difference between shortest distance of communities and the average shortest distance of each community
}
