% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v8.r
\name{link.gene.path}
\alias{link.gene.path}
\title{Plot shortest paths through a link gene between two modules}
\usage{
link.gene.path(g, x, y, linkgene, col1, col2, link.col)
}
\arguments{
\item{g}{Network data frame}

\item{linkgene}{Target link gene to visualize paths through}

\item{col1}{Color for first module genes}

\item{col2}{Color for second module genes}

\item{link.col}{Color for link genes}

\item{module1}{First module genes}

\item{module2}{Second module genes}
}
\description{
Plot shortest paths through a link gene between two modules
}
