% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoBCnet-package.R
\docType{package}
\name{MoBCnet-package}
\alias{MoBCnet}
\alias{MoBCnet-package}
\title{MoBCnet: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
