% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality.r
\name{plotMoBC_genes}
\alias{plotMoBC_genes}
\title{plotDist}
\usage{
plotMoBC_genes(
  MoBC.result,
  module1.name,
  module2.name,
  top = 10,
  module1.color = "lightblue1",
  module2.color = "lightpink"
)
}
\arguments{
\item{MoBC.result}{results from CommDistFunction function}

\item{module1.name}{The name of the community for which centrality is being calculated. This should be one of the communities provided as input}

\item{module2.name}{The name of the community for which centrality is being calculated. This should be one of the communities provided as input}

\item{top}{}

\item{module1.color}{}

\item{module2.color}{}
}
\value{
plot
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
plot.MoBC.genes(MoBC.result, module1.name, module2.name, 
                   top=10, module1.color='lightblue1',module2.color='lightpink')
}
