% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Degree_random_v2.r
\name{simple_sampling}
\alias{simple_sampling}
\title{module distance function}
\usage{
simple_sampling(deg, membership, ncv = 1000)
}
\description{
Calculate the distance and z_score of the communities
use.f can be choosed in the DistFunction.r
ex) get.shortest.dist, get.kernel.dist, get.centre.dist, get.separation.dist, get.closest.dist
As a default, get.closest.dist function is used to measure the distance between communities
User also can make the dist function and use it for calculating module distance
This function is made to know the z-score of a measured distance from distances of degree-preserved random networks
}
