% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v5.r
\name{cal.FCSgenes.values}
\alias{cal.FCSgenes.values}
\title{cal.FCSgenes.values}
\usage{
cal.FCSgenes.values(g, community1, community2, allg)
}
\arguments{
\item{g}{graph}

\item{community1}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}

\item{community2}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}
}
\value{
data.frame
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
cal.FCgene(graph, 'module_1','module_2')
}
