% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v3.r
\name{cal.MoBCgenes}
\alias{cal.MoBCgenes}
\title{cal.MoBCgenes}
\usage{
cal.MoBCgenes(g, community1, community2, random, ratio, cal.p, nCore)
}
\arguments{
\item{g}{graph}

\item{community1}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}

\item{community2}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}
}
\value{
vector
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
cal.MoBCgenes(graph, 'community1','community2',random,ratio,cal.p, nCore)
}
