% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v4.r
\name{cal.MoBCgenes}
\alias{cal.MoBCgenes}
\title{cal.MoBCgenes}
\usage{
cal.MoBCgenes(
  g,
  comm.genelist,
  community1n,
  community2n,
  random,
  ratio,
  randomMethod,
  nCore
)
}
\arguments{
\item{g}{graph}

\item{comm.genelist}{list of community genes}

\item{community1n}{The name of the community for which centrality is being calculated. This should be one of the communities provided as input}

\item{community2n}{The name of the community for which centrality is being calculated. This should be one of the communities provided as input}
}
\value{
vector
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
cal.MoBCgenes(graph, 'community1','community2',random,ratio,randomMethod, nCore)
}
