% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v4.r
\name{cal.MoBCgenes.values}
\alias{cal.MoBCgenes.values}
\title{cal.MoBCgenes.values}
\usage{
cal.MoBCgenes.values(g, community1, community2, allg)
}
\arguments{
\item{g}{graph}

\item{community1}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}

\item{community2}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}
}
\value{
vector
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
cal.MoBCgenes.values(graph, 'module_1','module_2', allg)
}
