% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistance_v3.r
\name{CommuinityDistance}
\alias{CommuinityDistance}
\title{module distance function}
\usage{
CommuinityDistance(
  network,
  module.genelist,
  randomMethod = c("None", "RandSD", "RandSDM"),
  random = 1000,
  ratio = 0.1,
  nCore = 1,
  method = c("closest", "shortest", "kernel", "centre", "separation")
)
}
\arguments{
\item{network}{A data frame representing the input biological network.}

\item{module.genelist}{A list of gene sets corresponding to predefined network modules.}

\item{randomMethod}{A character string specifying the method used to generate randomized networks.}

\item{random}{An integer indicating the number of randomized network generations and distance calculations to perform.}

\item{ratio}{A numeric value specifying the allowable ratio of overlap when sampling or generating randomized modules.}

\item{nCore}{An integer specifying the number of CPU cores to use for parallel computation.}

\item{method}{A character string specifying the distance measurement method to use; one of
\code{"closest"}, \code{"shortest"}, \code{"kernel"}, \code{"centre"}, or \code{"separation"}.}
}
\value{
A list containing the computed distance results, the input \code{module.genelist}, and the input \code{network}.
}
\description{
Calculate the distance and z_score of the communities
use.f can be choosed in the DistFunction.r
ex) get.shortest.dist, get.kernel.dist, get.centre.dist, get.separation.dist, get.closest.dist
As a default, get.closest.dist function is used to measure the distance between communities
User also can make the dist function and use it for calculating module distance
This function is made to know the z-score of a measured distance from distances of degree-preserved random networks
}
