% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistFunction.r
\name{get.shortest.dist}
\alias{get.shortest.dist}
\title{Distance Function}
\usage{
get.shortest.dist(distm, x, y)
}
\arguments{
\item{distm}{input}

\item{x}{input (gene cluster 1)}

\item{y}{input (gene cluster 2)

get.shortest.dist
mean of shortest path lengths of all pariwised community genes}
}
\description{
Calculate the distance through the dist matrix
}
