% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v4.r
\name{link.gene.path}
\alias{link.gene.path}
\title{Plot shortest paths through a link gene between two modules}
\usage{
link.gene.path(g, x, y, linkgene)
}
\arguments{
\item{g}{Network data frame}

\item{linkgene}{Target link gene to visualize paths through}

\item{module1}{First module genes}

\item{module2}{Second module genes}
}
\description{
Plot shortest paths through a link gene between two modules
}
