% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v3.r
\name{MoBC.genes}
\alias{MoBC.genes}
\title{Get.Centrality}
\usage{
MoBC.genes(
  network,
  module1.gene,
  module2.gene,
  randomMethod = c("None", "RandC", "RandCD", "RandCM", "RandCDM"),
  random = 1000,
  nCore = 1,
  ratio = 0.1
)
}
\arguments{
\item{network}{results from CommDistFunction function}

\item{module1.gene}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}

\item{module2.gene}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}
}
\value{
data.frame
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
Get.Centrality(MoBC.result, 'module_1','module_2')
}
