% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v8.r
\name{MoBC.genes}
\alias{MoBC.genes}
\title{MoBC.genes}
\usage{
MoBC.genes(
  network,
  module.gene.list,
  module1,
  module2,
  randomMethod = c("None", "RandSD", "RandSDM"),
  random = 1000,
  nCore = 1,
  ratio = 0.1
)
}
\arguments{
\item{network}{results from CommDistFunction function}

\item{module1}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}

\item{module2}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}
}
\value{
data.frame
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
MoBC.genes(network, module.gene.list, module1, module2)
}
