% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v4.r
\name{MoBC.genes}
\alias{MoBC.genes}
\title{MoBC.genes}
\usage{
MoBC.genes(
  network,
  module.gene.list,
  module1,
  module2,
  randomMethod = c("None", "RandSD", "RandSDM"),
  random = 1000,
  nCore = 1,
  ratio = 0.1
)
}
\arguments{
\item{network}{A data frame representing the input biological network, where rows define interactions between genes.}

\item{module.gene.list}{A list of gene sets corresponding to predefined network modules.}

\item{module1}{A character string specifying the first module for which MoBC-based centrality is calculated.
This must match one of the module names provided in \code{module.gene.list}.}

\item{module2}{A character string specifying the second module for which MoBC-based centrality is calculated.
This must match one of the module names provided in \code{module.gene.list}.}

\item{randomMethod}{A character string specifying the method used to generate randomized networks for significance assessment.}

\item{random}{An integer specifying the number of randomized network generations used to estimate centrality significance.}

\item{nCore}{An integer specifying the number of CPU cores to use for parallel computation.}

\item{ratio}{A numeric value indicating the maximum allowed overlap ratio when sampling or generating randomized modules.}
}
\value{
A data frame containing MoBC-based centrality scores and associated statistics for genes linking the two modules.
}
\description{
Calculate link gene centrality between two modules
}
