% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v3.r
\name{plotDist}
\alias{plotDist}
\title{plotDist}
\usage{
plotDist(MoBC.result, pval = 0.05)
}
\arguments{
\item{MoBC.result}{results from CommDistFunction function}

\item{pval}{cut-off for filtering edges between communities}

\item{module1.name}{The name of the community for which centrality is being calculated. This should be one of the communities provided as input}

\item{module2.name}{The name of the community for which centrality is being calculated. This should be one of the communities provided as input}

\item{top}{}

\item{module1.color}{}

\item{module2.color}{}
}
\value{
plot

plot
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
plot.MoBC.genes(MoBC.result, module1.name, module2.name, 
                   top=10, module1.color='lightblue1',module2.color='lightpink')
Calculate centrality between two modules from MoBC result 


plotDist(MoBC.result, pval=0.05)
}
