% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality_v4.r
\name{plotDist}
\alias{plotDist}
\title{plotDist}
\usage{
plotDist(MoBC.result, pval = 0.05)
}
\arguments{
\item{MoBC.result}{results from CommDistFunction function}

\item{pval}{cut-off for filtering edges between communities}
}
\value{
plot
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
plotDist(MoBC.result, pval=0.05)
}
